third_party/blink/web_tests/platform/generic/editing/inserting/insert-without-inheriting-style-expected.txt

PASS $("sample").innerHTML is "foo quux&nbsp;baz"
PASS successfullyParsed is true

TEST COMPLETE
